requirements.txt
# 原有依赖...
flask==2.3.3
gunicorn==21.2.0
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.3.0
joblib==1.3.2
mlflow==2.9.0
dvc==3.34.0
dagshub==0.6.3
uvicorn==0.24.0  # 新增uvicorn，用于启动Flask服务
gitpython==3.1.40  # train.py依赖（获取Git commit）
pytest==7.4.3  # 测试依赖
requests==2.31.0  # 测试时发送HTTP请求